Replace all Xom httpget()s with Stage httpget()s
It turns out that there are some instances where Xom.[async_]httpget is
being called from outside the MirrorStage, notably when serving packages
from PyPIView. These were therefore not getting auth credentials added
to them and receiving a 403 from the remote. For these cases the stage
is also available when the function is being called, so the simple
solution is simply to replace the xom.httpget() call with a
corresponding stage.httpget() call.